Cleaner binary compatibility solution
